requirements.txt
anyio==3.7.1    
certifi==2023.7.22
charset-normalizer==3.2.0    
class-registry==2.1.2    
python-telegram-bot==20.4
pytz==2023.3
regex==2023.6.3
requests==2.31.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
tabula-py==2.7.0
telegram==0.0.1
tzdata==2023.3
urllib3==2.0.4
